Probe dashboard
Add a simple dashboard view to all probes with some charts and tables.
The configuration is done in the event classes.

This is not meant to be a replacement for tools like Kibana or Grafana.
This was built as an experiment to give the users a quick overview, and
hopefully save them some time.