Fix monolith enrollment packages
The previous commit ce338c7fd8c introduced a bug. It is not possible to
build the enrollment package when they are downloaded if they are
missing. Their checksum would be different than the one present in the
pkg_info when the catalog was built, and munki would reject them. (The
checksum is updated by build_manifest_enrollment_package, but it is too
late.)

The rebuild_manifest_enrollment_packages command should be used instead,
if some packages are missing.

We also fix a bug in the osquery enrollment package. When we added the
option to disable or enable the file carver, we forgot the previously
configured osquery enrollment packages which didn't have the option. We
set it to "disabled" per default, like in the enrollment package form.