Better password validation and session management
Django password validators can be used now with the
AUTH_PASSWORD_VALIDATORS variable in base.json

Password history is saved.

MAX_PASSWORD_AGE_DAYS can be used in the django section of base.json to
force a password change after a certain number of days.

accounts.password_validation.PasswordNotAlreadyUsedValidator can be used
as password validator to check that the passwords are unique per user.

The SESSION_COOKIE_AGE and SESSION_EXPIRE_AT_BROWSER_CLOSE variables are
copied from the django section of base.json into the django settings, if
present. This enables a better control of the session validity.