MDM OTA Enrollment
A profile service payload can be constructed to enroll devices in the
Zentral MDM, which cannot do anything at the moment. This is the bare
minimum to be able to authorize kernel extensions in macOS 10.13.

This prepares also the use of client certificates for the agents running
on the managed machines (osquery, santa, …). A SCEP server
(github.com/micromdm/scep) is included, and the CSR are verified by
Zentral.