Fix product archive title bug
get_product_archive_title() was used to find out if a package was a
product archive. Not anymore.

A package is a product archive if it is based on a product archive, or
if extra packages are provided. If a package is a product archive, we
use the get_product_archive_title() method to get its title.