Split domains for client cert authentication
There are a lot of problems with the optional client certificate
authentication on a single domain. As soon as a valid client cert for
the domain in present in the keychain (MDM for example), it will be used
by santa, or safari.
With this commit, we undo the recent work, but we try to keep only one
certificate, using the Subject Alternative Name extension. This way,
only one chain has to be distributed to the clients.