Jamf instance info → observer event metadata
The jamf instance info were part of the event payload. Because the
elasticsearch event store uses the event type as a namespace for the
event attributes, the jamf instance info could not be simply extracted
when working with more than one type of event.
By moving the jamf instance info into a metadata attribute, we make sure
that they are not namespaced by the elasticsearch store.
We use the ECS observer fields.