Add runworker command and some runworkers fixes
runworker command can be used to list all the workers, and run a single
worker, without multiprocessing and lifecycle management. Useful to let
the OS manage the worker, or run a single worker in a container.

Both commands now accept a --json arg to serialize the --list-workers
output in JSON. This can be used in setup scripts for example.

The --list-workers options of runworkers doesn't hang anymore if they
are no workers to launch.