Fix probe sync bug when no probes
The sync thread would stop if there were no probes in the probe view,
because a probe view evaluates to False if empty. We now test if the
resolved weak reference is not None.

The liveness of the sync thread is also tested everytime the _load
method of a probe view is called, and a new thread is started if the
old one is dead – belt-and-suspenders.