Better solution for custom error pages
Add build_custom_error_pages management command to render correct html
error pages that can be used by the HTTP proxy, when the application is
failing.
Remove buggy nginx error_page docker configurations. The new pages are
not configured, because they need to be built before they can be used.
It could be done in the docker-entrypoint, but the web app start is
already slow enough. PRs welcomed!