Report munki full apps info & locally save tags
Add `apps_full_info_shard` for the munki app. This value if present is
sent to the postflight script, and will be use to decide if the full
application info (std + bundle display name and bundle version) will be
uploaded. If the option is absent, those info will not be uploaded. A
value between 1 and 100 will only activate this on some machines (%).

The job details API call has been moved to the preflight. The job
details are temporarily saved in the `/usr/local/zentral/munki` dir for
the postflight (to avoid a second API call). The preflight script looks
for the `tags` key in the job details and save them as `zentral_tags` in
the default `ConditionalItems.plist`. It is done in the preflight to
allow the tags to be used as soon as possible in during a munki run.