Add proxied object to interact with the stores
Also add `events_url_authorized_groups` common store configuration key,
to set the list of group names whose members can be redirected to the
store. If not set, the events URLs will be displayed to all users. This
is used only if the store itself implements the requested events URLs.

The new Stores object has methods to filter the stores that implement
the requested events URLs and are authorized for a given user.