Reorganize the machine update events
Replace the inventory update events with machine update events. Use the
same namespace "machine". This way, we maintain the same fields in the
events store accross the events. We also solve the "source" attribute
conflict. Finally, we separate the Add from the Remove event. This is to
facilitate the search in all the stores. This is possible now that we
are using a common namespace, without doubling the number of fields.