RBAC on the nginx auth endpoint
Verify the first path element of the original URL received from Nginx
and compare it to the configured extra_links. If there is a match, and
the extra_link is configured with `authorized_groups` (list of group
names whose members are authorized to access the link), the access is
granted.
**WARNING**
If there is no match or there are no `authorized_groups` the access is
granted too!!! (Might change in the future, this is the legacy
behaviour.)