Use "requires" for monolith enrol. packages
BREAKING CHANGE for monolith module: you need to update your manifest
enrollment packages to use "requires" instead of "update_for".

Avoid requesting removal from packages, like "osquery", when no
enrollment package is in scope for a machine. This way, we fix an issue
when "osquery" is a managed install in some sub manifest.