Add AWS SNS/SQS use_filter_policies option
If set to true, the AWS SNS/SQS queues backend will set filter policies
on the subscriptions used for the stores queues, if the stores are
configured with "included_event_types" or "excluded_event_types". This
way, only the events with the right type will be forwarded to the
store workers.