Fix MDM blueprint API concurrency issue
The blueprint serialized artifacts must be updated after the artifacts,
profiles, and blueprint artifacts change. If changes occur concurrently,
the serialized artifacts may be inconsistent, because of the transaction
isolation in PostgreSQL. We use multiple transactions and locks on the
blueprint to make sure all changes are seen before the serialized
artifacts are updated.