Ensure Santa ballot unicity across realms
Make sure that not only a ballot from the same realm user is a match,
but also a ballot from a different realm user with the same username.
This avoids double votes with 2 SAML apps with different scopes (All the
employees for the portal, only the admins for the Zentral console) but
from the same IdP, so with the same usernames.